[kafka_actions] Break formats <-> message_deserializer import cycle
Move the standalone deserialization helpers (_deserialize_*/_build_* plus the
protobuf index/class helpers and the well-known-type preloader) into a new
formats/_helpers.py module. builtins.py and extras.py now import them from
formats._helpers, and message_deserializer.py no longer defines them, so the
dependency flows one way (message_deserializer -> formats) with no cycle.
