+++ kafka_actions/datadog_checks/kafka_actions/formats/builtins.py
@@ -3,14 +3,14 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 """Built-in format handlers.
 
-Thin wrappers over the module-level helpers in
-``..message_deserializer`` so deserialization logic lives in one place.
+Thin wrappers over the standalone helpers in :mod:`._helpers` so the
+deserialization logic lives in one place and the dependency flows one way
+(``message_deserializer`` -> ``formats``).
 """
 
 from __future__ import annotations
 
-from datadog_checks.kafka_actions.formats.base import FormatHandler
-from datadog_checks.kafka_actions.message_deserializer import (
+from datadog_checks.kafka_actions.formats._helpers import (
     _build_avro_schema,
     _build_protobuf_schema,
     _build_protobuf_schema_from_registry,
@@ -21,6 +21,7 @@ from datadog_checks.kafka_actions.message_deserializer import (
     _deserialize_raw,
     _deserialize_string,
 )
+from datadog_checks.kafka_actions.formats.base import FormatHandler
 
 
 class JsonHandler(FormatHandler):
