+++ kafka_actions/datadog_checks/kafka_actions/message_deserializer.py
@@ -15,25 +15,8 @@ Additional handlers/codecs may be registered via the entry-point groups.
 from __future__ import annotations
 
 import base64
-import datetime
-import decimal
 import hashlib
 import json
-import uuid
-
-from google.protobuf import (
-    any_pb2,
-    api_pb2,
-    descriptor_pb2,
-    duration_pb2,
-    empty_pb2,
-    field_mask_pb2,
-    source_context_pb2,
-    struct_pb2,
-    timestamp_pb2,
-    type_pb2,
-    wrappers_pb2,
-)
 
 from .compression import get_codec as _get_compression_codec
 from .compression.registry import register_codec as _register_codec
@@ -45,229 +28,6 @@ SCHEMA_REGISTRY_MAGIC_BYTE = 0x00
 # Distinct sentinel so a legitimately cached None (schemaless handlers) is not
 # treated as a cache miss and rebuilt on every message.
 MISSING_SCHEMA = object()
-
-_WELL_KNOWN_TYPE_MODULES = (
-    any_pb2,
-    duration_pb2,
-    empty_pb2,
-    field_mask_pb2,
-    source_context_pb2,
-    struct_pb2,
-    timestamp_pb2,
-    wrappers_pb2,
-    type_pb2,
-    api_pb2,
-)
-
-
-def _preload_well_known_types(pool):
-    """Add google/protobuf/*.proto well-known types to a fresh DescriptorPool."""
-    for module in _WELL_KNOWN_TYPE_MODULES:
-        file_name = module.DESCRIPTOR.name
-        try:
-            pool.FindFileByName(file_name)
-            continue
-        except KeyError:
-            pass
-        fd_proto = descriptor_pb2.FileDescriptorProto()
-        module.DESCRIPTOR.CopyToProto(fd_proto)
-        pool.Add(fd_proto)
-
-
-class _AvroJSONEncoder(json.JSONEncoder):
-    """JSON encoder that handles types returned by fastavro for Avro logical types."""
-
-    def default(self, obj):
-        if isinstance(obj, decimal.Decimal):
-            return str(obj)
-        if isinstance(obj, (datetime.datetime, datetime.date, datetime.time)):
-            return obj.isoformat()
-        if isinstance(obj, uuid.UUID):
-            return str(obj)
-        if isinstance(obj, bytes):
-            return base64.b64encode(obj).decode('ascii')
-        return super().default(obj)
-
-
-def _read_varint(data):
-    shift = 0
-    result = 0
-    bytes_read = 0
-    for byte in data:
-        bytes_read += 1
-        result |= (byte & 0x7F) << shift
-        if (byte & 0x80) == 0:
-            return result, bytes_read
-        shift += 7
-    raise ValueError("Incomplete varint")
-
-
-def read_protobuf_message_indices(payload):
-    """Read the Confluent Protobuf message indices array after the schema ID."""
-    array_len, bytes_read = _read_varint(payload)
-    payload = payload[bytes_read:]
-    indices = []
-    for _ in range(array_len):
-        index, bytes_read = _read_varint(payload)
-        indices.append(index)
-        payload = payload[bytes_read:]
-    return indices, payload
-
-
-def get_protobuf_message_class(schema_info, message_indices):
-    from google.protobuf import message_factory
-
-    pool, descriptor_set = schema_info
-    file_descriptor = descriptor_set.file[0]
-    message_descriptor_proto = file_descriptor.message_type[message_indices[0]]
-    package = file_descriptor.package
-    name_parts = [message_descriptor_proto.name]
-
-    current_proto = message_descriptor_proto
-    for idx in message_indices[1:]:
-        current_proto = current_proto.nested_type[idx]
-        name_parts.append(current_proto.name)
-
-    full_name = f"{package}.{'.'.join(name_parts)}" if package else '.'.join(name_parts)
-    message_descriptor = pool.FindMessageTypeByName(full_name)
-    return message_factory.GetMessageClass(message_descriptor)
-
-
-def _deserialize_json(message):
-    if not message:
-        return None
-    decoded = message.decode('utf-8').strip()
-    if not decoded:
-        return None
-    json.loads(decoded)
-    return decoded
-
-
-def _deserialize_string(message):
-    if not message:
-        return None
-    decoded = message.decode('utf-8')
-    if not decoded:
-        return None
-    return json.dumps(decoded)
-
-
-def _deserialize_raw(message):
-    if not message:
-        return None
-    return json.dumps(base64.b64encode(message).decode('ascii'))
-
-
-def _deserialize_bson(message):
-    if not message:
-        return None
-    from bson import decode as bson_decode
-    from bson.json_util import dumps as bson_dumps
-
-    try:
-        return bson_dumps(bson_decode(message))
-    except Exception as e:
-        raise ValueError(f"Failed to deserialize BSON message: {e}") from e
-
-
-def _deserialize_avro(message, schema):
-    if not message:
-        return None
-    if schema is None:
-        raise ValueError("Avro schema is required")
-    from io import BytesIO
-
-    from fastavro import schemaless_reader
-
-    try:
-        bio = BytesIO(message)
-        initial = bio.tell()
-        data = schemaless_reader(bio, schema)
-        bytes_read = bio.tell() - initial
-        if bytes_read != len(message):
-            raise ValueError(
-                f"Not all bytes were consumed during Avro decoding! "
-                f"Read {bytes_read} bytes, but message has {len(message)} bytes."
-            )
-        return json.dumps(data, cls=_AvroJSONEncoder)
-    except Exception as e:
-        raise ValueError(f"Failed to deserialize Avro message: {e}") from e
-
-
-def _deserialize_protobuf(message, schema, uses_schema_registry):
-    from google.protobuf.json_format import MessageToJson
-
-    if not message:
-        return None
-    if schema is None:
-        raise ValueError("Protobuf schema is required")
-    try:
-        if uses_schema_registry:
-            message_indices, message = read_protobuf_message_indices(message)
-            if not message_indices:
-                message_indices = [0]
-        else:
-            message_indices = [0]
-
-        message_class = get_protobuf_message_class(schema, message_indices)
-        instance = message_class()
-        consumed = instance.ParseFromString(message)
-        if consumed != len(message):
-            raise ValueError(
-                f"Not all bytes were consumed during Protobuf decoding! "
-                f"Read {consumed} bytes, but message has {len(message)} bytes."
-            )
-        return MessageToJson(instance)
-    except Exception as e:
-        raise ValueError(f"Failed to deserialize Protobuf message: {e}") from e
-
-
-def _build_avro_schema(schema_str):
-    schema = json.loads(schema_str)
-    if schema is None:
-        raise ValueError("Avro schema cannot be None")
-    return schema
-
-
-def _build_protobuf_schema(schema_str):
-    from google.protobuf import descriptor_pb2, descriptor_pool
-
-    schema_bytes = base64.b64decode(schema_str)
-    descriptor_set = descriptor_pb2.FileDescriptorSet()
-    descriptor_set.ParseFromString(schema_bytes)
-
-    pool = descriptor_pool.DescriptorPool()
-    _preload_well_known_types(pool)
-    for fd_proto in descriptor_set.file:
-        pool.Add(fd_proto)
-    return (pool, descriptor_set)
-
-
-def _build_protobuf_schema_from_registry(schema_str, dep_schemas):
-    from google.protobuf import descriptor_pb2, descriptor_pool
-
-    pool = descriptor_pool.DescriptorPool()
-    _preload_well_known_types(pool)
-    descriptor_set = descriptor_pb2.FileDescriptorSet()
-
-    for dep_name, dep_b64 in dep_schemas:
-        try:
-            pool.FindFileByName(dep_name)
-            continue
-        except KeyError:
-            pass
-        dep_bytes = base64.b64decode(dep_b64)
-        dep_proto = descriptor_pb2.FileDescriptorProto()
-        dep_proto.ParseFromString(dep_bytes)
-        dep_proto.name = dep_name
-        pool.Add(dep_proto)
-
-    schema_bytes = base64.b64decode(schema_str)
-    fd_proto = descriptor_pb2.FileDescriptorProto()
-    fd_proto.ParseFromString(schema_bytes)
-    descriptor_set.file.append(fd_proto)
-    pool.Add(fd_proto)
-    return (pool, descriptor_set)
 
 
 def _bootstrap_format_handlers():
